requirements.txt
antigate==1.4.0
EasyProcess==0.2.2
Pillow==3.2.0
PyVirtualDisplay==0.2
selenium==2.53.1
six==1.10.0
wheel==0.24.0
xmltodict==0.10.1
